move file size to Common component
